fix(portao): os DOIS vermelhos de clippy do fecho -- e o segundo fica MAIS FORTE, nao silenciado
1. `b.port as usize` onde `port` JA' e' `usize` (`unnecessary_cast`) -- meu, do commit 47cf9d2bd.

2. `assert!(MAX_DEMO_LEVEL >= 119)` e' uma assercao entre duas CONSTANTES, e o clippy tem razao:
   ela decide-se antes de o teste correr. ⇒ parte em DUAS metades que afirmam coisas diferentes:
   um `const _: () = assert!(...)` (baixar o tecto deixa de COMPILAR, que e' mais forte que
   reprovar) e uma que MEDE a tabela das cenas de ciclo (`is_cycle_scene("119")`, mais a metade
   da obsolescencia: a tabela nao pode responder por uma cena que nao existe).
   ⚠️ Um tecto alto com a tabela sem a linha abre o canvas em BRANCO -- sao dois factos, e o
   `assert!` antigo so' tocava no primeiro.

+++ crates/ph2d-motion-diagnose/src/lib_tests.rs
@@ -103,7 +103,7 @@ fn the_transient_set_covers_every_consumed_column() {
             }
             // O tipo que este binding LÊ contra o que o nó EMITE. Uma porta fora de alcance
             // não isenta ninguém: o `map_or(false, …)` deixa a asserção decidir.
-            let lido = m.inputs.get(b.port as usize).map(|p| p.ty);
+            let lido = m.inputs.get(b.port).map(|p| p.ty);
             let emitido = m.outputs.first().map(|p| p.ty);
             if lido.is_some() && emitido.is_some() && lido != emitido {
                 transdutores += 1;
